+++ tests/cli_pl.rs
@@ -63,7 +63,6 @@ fn command_overlap() -> anyhow::Result<()> {
     Ok(())
 }
 
-
 #[test]
 fn command_orient() -> anyhow::Result<()> {
     let mut bin = String::new();
@@ -92,6 +91,35 @@ fn command_orient() -> anyhow::Result<()> {
 
     assert_eq!(stdout.lines().count(), 24);
     assert!(stdout.contains("pac4745_7148"), "original names");
+
+    Ok(())
+}
+
+#[test]
+fn command_contained() -> anyhow::Result<()> {
+    let mut bin = String::new();
+    for e in &["LAshow"] {
+        if let Ok(pth) = which::which(e) {
+            bin = pth.to_string_lossy().to_string();
+            break;
+        }
+    }
+    if bin.is_empty() {
+        return Ok(());
+    } else {
+        eprintln!("bin = {:#?}", bin);
+    }
+
+    let mut cmd = Command::cargo_bin("anchr")?;
+    let output = cmd
+        .arg("contained")
+        .arg("tests/ovlpr/contained.fasta")
+        .output()
+        .unwrap();
+    let stdout = String::from_utf8(output.stdout).unwrap();
+
+    assert_eq!(stdout.lines().count(), 2);
+    assert!(stdout.contains("infile_0/4/0_15361"), "renamed");
 
     Ok(())
 }