+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add `anchr show2ovlp`
 * Add `anchr overlap`
 * Add `anchr orient`
+* Add `anchr contained`
 * Move `ovlpr` subcommands from intspan here
     * `anchr covered`
     * `anchr paf2ovlp`